style Playlist Header with styled components
